agregando reporte

+++ app/Http/Controllers/EmpaqueController.php
@@ -2,12 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\EditEmpaqueRequest;
 use App\Http\Requests\EmpaqueRequest;
 use App\Models\Almacen;
 use App\Models\Empaque;
 use App\Models\Empresa;
 use App\Models\ListaEmpaques;
+use App\Models\Movimiento;
 use App\Models\UbicacionAlmacen;
+use DB;
 use Illuminate\Http\Request;
 
 class EmpaqueController extends Controller
@@ -23,19 +26,16 @@ class EmpaqueController extends Controller
             ->orderBy('empaque.id', 'desc')
             ->get();
         
-        /*
-        $empaques = Empaque::leftJoin('lista_empaques', 'empaque.lista_empaques_id', '=', 'lista_empaques.id')
-        ->leftJoin('ubicacion_almacen', 'empaque.ubicacion_almacen_id', '=', 'ubicacion_almacen.id')
-        ->leftJoin('almacen', 'ubicacion_almacen.almacen_id', '=', 'almacen.id')
-        ->select(
-            'empaque.*', 
-            'lista_empaques.codigo as empaque_lista_empaque_codigo', 
-            'almacen.nombre as empaque_almacen_nombre', 
-            'ubicacion_almacen.nombre as empaque_ubicacion_nombre'
-        )
-        ->orderBy('empaque.id', 'desc')
-        ->get();
-        */
+        foreach($empaques as $empaque){
+            $movimientos = Movimiento::where('empaque_id', $empaque->id)->count();
+            if($movimientos !== null && $movimientos > 1){
+                $empaque->tiene_movimientos = true;
+            }else{
+                $empaque->tiene_movimientos = false;
+            }
+        }
+        
+        
 
         $almacenes = Almacen::all();
         foreach ($almacenes as $almacen) {
@@ -44,7 +44,7 @@ class EmpaqueController extends Controller
                 ->get();
         }
 
-        $empresa = Empresa::find(1);
+        $empresa = Empresa::find(auth()->user()->empresa_id);
         $icono_empresa = $empresa->icono;
 
         $listas = ListaEmpaques::all();
@@ -88,11 +88,36 @@ class EmpaqueController extends Controller
         return redirect()->route('lista_empaques.index');
     }
 
+    public function update($id, EditEmpaqueRequest $request)
+    {
+        $empaque = Empaque::findOrFail($id);
+        
+        $empaque->tipo = $request->tipo;
+        $empaque->cantidad_cajas = $request->cantidad_cajas;
+        $empaque->peso = $request->peso;
+        $empaque->unidad_medida = $request->unidad_medida;
+        $empaque->descripcion = $request->descripcion;
+        $empaque->estado = $request->estado;
+        $empaque->observacion_estado = $request->observacion_estado;
+        $empaque->criterio1 = $request->has('criterio1') ? true : false;
+        $empaque->criterio2 = $request->has('criterio2') ? true : false;
+        $empaque->criterio3 = $request->has('criterio3') ? true : false;
+        $empaque->update();
+
+        return redirect()->route('empaque.index')->with('success', 'Lista de empaques actualizada exitosamente.');
+    }
+
     public function delete($id)
     {
-        
-        $empaque = Empaque::findOrFail($id);
-        $empaque->delete();
-
+        $cantMovimientos = Movimiento::where('empaque_id',$id)->count();
+        if($cantMovimientos == null || $cantMovimientos <2){
+            $empaque = Empaque::findOrFail($id);
+                        $empaque->delete();
+                        $listaEmpaque = ListaEmpaques::find($empaque->lista_empaques_id);
+                        $listaEmpaque->stock_registrado = $listaEmpaque->stock_registrado-1;
+                        $listaEmpaque->stock_actual = $listaEmpaque->stock_actual-1;
+                        $listaEmpaque->update();
+        }
+        return redirect()->route('empaque.index');
     }
 }
